added utils.py
