+++ monet_boys/discriminator.py
@@ -21,68 +21,34 @@ will use instance normalization by default, but will optionally apply dropout (t
 
 # Global variables SUGGESTION: lets move sampling to utilities.py for cleaner code
 
-OUTPUT_CHANNELS = 3
+#def downsample(filters, size, apply_instancenorm=True):
+#    """ defines and returns a downsampler layer with parameters :
+#    filter (dimensionality of output space)
+#    size (kernel size)
+#    (+ optionally applies instance normalization)
+#    """
+#    #Initialize random distributions for the kernel initializer and the instance normalizer
+#    initializer = tf.random_normal_initializer(0., 0.02)
+#    gamma_init = keras.initializers.RandomNormal(mean=0.0, stddev=0.02)
+#
+#    # Instantiate model
+#    result = keras.Sequential()
+#
+#    # Add 2D convolutional layer w. stride 2 and padding
+#    result.add(layers.Conv2D(filters, size, strides=2, padding='same',
+#                             kernel_initializer=initializer, use_bias=False))
+#
+#    # Apply instance normalization
+#    if apply_instancenorm:
+#        result.add(tfa.layers.InstanceNormalization(
+#            gamma_initializer=gamma_init))
+#
+#    # Add a last Leaky ReLU layer
+#    result.add(layers.LeakyReLU())
+#
+#    return result
 
-
-def downsample(filters, size, apply_instancenorm=True):
-    """ defines and returns a downsampler layer with parameters :
-    filter (dimensionality of output space)
-    size (kernel size)
-    (+ optionally applies instance normalization)
-    """
-    #Initialize random distributions for the kernel initializer and the instance normalizer
-    initializer = tf.random_normal_initializer(0., 0.02)
-    gamma_init = keras.initializers.RandomNormal(mean=0.0, stddev=0.02)
-
-    # Instantiate model
-    result = keras.Sequential()
-
-    # Add 2D convolutional layer w. stride 2 and padding
-    result.add(layers.Conv2D(filters, size, strides=2, padding='same',
-                             kernel_initializer=initializer, use_bias=False))
-
-    # Apply instance normalization
-    if apply_instancenorm:
-        result.add(tfa.layers.InstanceNormalization(
-            gamma_initializer=gamma_init))
-
-    # Add a last Leaky ReLU layer
-    result.add(layers.LeakyReLU())
-
-    return result
-
-
-def upsample(filters, size, apply_dropout=False):
-    """ defines and returns an upsampler layer with parameters :
-        filter (dimensionality of output space)
-        size (kernel size)
-        (+ optionally applies droupout)
-    """
-    #Initialize random distributions for the kernel initializer and the instance normalizer
-    initializer = tf.random_normal_initializer(0., 0.02)
-    gamma_init = keras.initializers.RandomNormal(mean=0.0, stddev=0.02)
-
-    # Instantiate model
-    result = keras.Sequential()
-
-    # Add transposed 2D convolutional layer w. stride 2 and padding
-    result.add(layers.Conv2DTranspose(filters, size, strides=2,
-                                      padding='same',
-                                      kernel_initializer=initializer,
-                                      use_bias=False))
-
-    # Apply instance normalization
-    result.add(tfa.layers.InstanceNormalization(gamma_initializer=gamma_init))
-
-    # Apply droupout
-    if apply_dropout:
-        result.add(layers.Dropout(0.5))
-
-    # Add a last ReLU layer
-    result.add(layers.ReLU())
-
-    return result
-
+from utils import downsample
 
 def Discriminator():
     initializer = tf.random_normal_initializer(0., 0.02)
